@@ -137,11 +141,20 @@ allprojects {
 
 gradleProjects {
     subprojects {
-        dirs(['libs', 'examples']) { Project subproject ->
-            apply plugin: 'micronaut-compatibility'
+        dirs(['libs']) { Project subproject ->
+            java {
+                toolchain {
+                    languageVersion = JavaLanguageVersion.of(17)
+                }
+            }
 
-            sourceCompatibility = 1.8
-            targetCompatibility = 1.8
+            micronaut {
+                importMicronautPlatform = true
+                testRuntime 'spock'
+                processing {
+                    incremental false
+                }
+            }
 
             repositories {
                 mavenCentral()
