@@ -78,7 +78,7 @@ public @interface Job {
     String fixedRate() default "";
 
     /**
-     * @return The name of a {@link javax.inject.Named} bean that is a
+     * @return The name of a {@link jakarta.inject.Named} bean that is a
      * {@link java.util.concurrent.ScheduledExecutorService} to use to schedule the task
      */
     String scheduler() default TaskExecutors.SCHEDULED;
