@@ -18,41 +18,19 @@
 package com.agorapulse.worker.sqs.v1
 
 import com.agorapulse.worker.tck.queue.AbstractQueuesSpec
-import com.amazonaws.auth.AWSCredentialsProvider
-import com.amazonaws.services.sqs.AmazonSQS
-import com.amazonaws.services.sqs.AmazonSQSClientBuilder
-import io.micronaut.context.ApplicationContext
-import org.testcontainers.containers.localstack.LocalStackContainer
-import org.testcontainers.spock.Testcontainers
-import spock.lang.Shared
+import io.micronaut.context.annotation.Property
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
 
-@Testcontainers
+/**
+ * Tests for SQS queues.
+ */
+@MicronautTest(environments = QUEUE_SPEC_ENV_NAME)
+@Property(name = 'aws.sqs.auto-create-queue', value = 'true')
+@Property(name = 'worker.jobs.send-words-job-listen.enabled', value = 'true')
+@Property(name = 'worker.jobs.send-words-job-hello.enabled', value = 'true')
 class SqsQueuesSpec extends AbstractQueuesSpec {
 
     @SuppressWarnings('GetterMethodCouldBeProperty')
     Class<?> getExpectedImplementation() { return SqsQueues }
-
-    @Shared
-    LocalStackContainer localstack = new LocalStackContainer().withServices(LocalStackContainer.Service.SQS)
-
-    @Override
-    ApplicationContext buildContext(String[] envs) {
-        AmazonSQS sqs = AmazonSQSClientBuilder
-            .standard()
-            .withEndpointConfiguration(localstack.getEndpointConfiguration(LocalStackContainer.Service.SQS))
-            .withCredentials(localstack.defaultCredentialsProvider)
-            .build()
-
-        return ApplicationContext
-            .builder(envs)
-            .properties(
-                'aws.sqs.auto-create-queue': 'true',
-                'worker.jobs.send-words-job-listen.enabled': 'true',
-                'worker.jobs.send-words-job-hello.enabled': 'true'
-            )
-            .build()
-            .registerSingleton(AmazonSQS, sqs)
-            .registerSingleton(AWSCredentialsProvider, localstack.defaultCredentialsProvider)
-    }
 
 }
